Create Day-1-3-exercise
Write a program that prints the number of characters in a user's name. You might need to Google for a function that calculates the length of a string.
e.g.

https://www.google.com/search?q=how+to+get+the+length+of+a+string+in+python+stack+overflow

Warning. Your program should work for different inputs. e.g. any name that you input.

Example Input
Aiswarya

Example Output

8