Create Project - Tip-calculator
Let's say that we and a couple of friends went out for lunch and it came to 124.56 dollars. And then it'll ask you, what percentage tip would you like to give? So let's say we want to give a 12 percent tip. So we enter 12 and hit enter. And then finally it asks how many people are splitting this bill? Let's say there are seven of us and then hit enter.
Finally, it's going to tell us that each person should pay about $19.93.

#If the bill was $150.00, split between 5 people, with a 12% tip.
#Each person should pay (150.00 / 5) * 1.12 = 33.6
#1.12 is (12/100)+1
#Format the result to 2 decimal places = 33.60
#Tip: There are 2 ways to round a number. You might have to do some Googling to solve this.💪
#HINT 1: https://www.google.com/search?q=how+to+round+number+to+2+decimal+places+python&oq=how+to+round+number+to+2+decimal
#HINT 2: https://www.kite.com/python/answers/how-to-limit-a-float-to-two-decimal-places-in-python
https://docs.python.org/3/tutorial/floatingpoint.html

Questions to be asked:
What was the total bill?
What percentage tip would you like to give? 10, 12, or 15?
How many people to split the bill?
Output should be
Each person should pay: Final_amount