Create Project - Rock Paper Scissors
Rock Paper Scissors:

https://www.wrpsa.com/

The game starts out by asking you to type 0 for rock, 1 for paper, or 2 for scissors. So let's go ahead and type two for scissors, and then it's going to show you a graphic of your choice scissors, and then the choice the computer made, which is rock. And of course, rock beats scissors so you lose. What this game boils down to is some way of randomly making a choice between rock, paper, or scissors and then comparing that choice that the computer randomly generated against your choice. And then based on the rules of rock, paper, and scissors, determining whether if you won, you lost or whether if you had a draw.