Create Project - Password-generator
Password-generator

Our program will ask us how many letters would you like in your password? 
So I will say I would like a long password, let's say 14.
How many symbols, 3, how many numbers, 4, and it's going to generate me a password that is almost completely impossible to crack by any hacker, no matter how talented they are.

Now there's two levels to this project.
You can either go for the easy version where all you have to do is generate your password in sequence. So for example, if the user wanted four letters, two symbols, and two numbers, then it might just look like this four letters and then two symbols and then two numbers. So they're in sequence right? First, you've got the letters, then you've got the number of required symbols, and then you've got the required numbers. So this is the easy level. And then there is the hard level. And the hard level requires you to generate the password in completely random order. So instead of having it being, letter symbol number, it should be maybe letter and then a symbol and then a number, then a letter, then a symbol, and then a number. But notice how this is completely random. Every single time you generate it, the order of where the symbols and where the numbers appear in your password are completely random. So for example, if you look at the demo version and I do the same thing, four letters, two symbols, two numbers, where the numbers and where the symbols occur are completely unpredictable. And every single time I generate a new password, it's going to switch it up again. So that's the hard level. So the project of today is to build this password generator.